Created getTotalQuantity and handleToast functions in lib and used handleToast in cart. #10
