+++ client/src/lib/get-total-quantity.js
@@ -1,0 +1,3 @@
+export default function getTotalQuantity(cartInventory) {
+  return cartInventory.reduce((total, item) => total + item.quantity, 0);
+}
