+++ client/src/lib/index.js
@@ -1,3 +1,5 @@
 export * from './api';
 export { default as toDollars } from './to-dollars';
 export { default as createLineBreaks } from './create-line-breaks';
+export { default as handleToast } from './handle-toast' ;
+export { default as getTotalQuantity} from './get-total-quantity';
